added new file validations.rb moved, 3 validations to new file from main.rb
